lesson-1/3-fixed-chatbot/requirements.txt
galileo[openai]
python-dotenv